pyproject.toml
[project]
name = "nonebot-plugin-clovers"
version = "0.3.0"
description = ""
authors = [{ name = "KarisAya", email = "1048827424@qq.com" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic (>=2.11.4,<3.0.0)",
    "nonebot2 (>=2.4.2,<3.0.0)",
    "clovers(>=0.4.1,<0.5.0)",
]
[project.optional-dependencies]
nonebot_adapter_onebot = ["nonebot-adapter-onebot (>=2.4.6,<3.0.0)"]
nonebot_adapter_qq = ["nonebot-adapter-qq (>=1.6.1,<2.0.0)"]
nonebot_adapter_satori = ["nonebot-adapter-satori (>=0.13.3,<0.14.0)"]
all = [
    "nonebot-adapter-onebot (>=2.4.6,<3.0.0)",
    "nonebot-adapter-qq (>=1.6.1,<2.0.0)",
    "nonebot-adapter-satori (>=0.13.3,<0.14.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
